fix(theming): retain the QStyle reference, plus review 3.6 cleanups
QWidget.setStyle() and QApplication.setStyle() do not take ownership of the
QStyle returned by QStyleFactory.create(), so DockThemeBridge handed the
target a style object with no remaining Python reference and let it be
collected. Keep it on the bridge. The isinstance(QApplication) branch there
was dead - both arms called self._target.setStyle(style).

Other 3.6 items:
- Document get()/get_all() as returning the live QColor objects, not copies;
  mutating one corrupts the theme process-wide. Copying was rejected
  deliberately: this sits on the per-widget refresh path.
- except (ValueError, TypeError, Exception) is just except Exception.
- ThemeJson.title_margin and .indicator_width were Optional[float] where
  ThemeSpec declares int. Now int. Behaviour change: a JSON theme with a
  fractional value (title_margin: 4.5) is rejected with a clear validation
  error instead of silently flowing into an int-declared field. Integral
  floats (4.0) still coerce.
- Document that a QSS theme overrides the palette engine and can leave the
  painted dock chrome half-applied.

Bump version to 0.5.16.

+++ lace/theme_models.py
@@ -126,7 +126,9 @@ class ThemeJson(BaseModel):
     title_padding_left: Optional[int] = None
     title_padding_right: Optional[int] = None
     title_button_spacing: Optional[int] = None
-    title_margin: Optional[float] = None
+    # int, not float — ThemeSpec.title_margin is int. Kept aligned so the JSON
+    # schema and the dataclass cannot drift apart.
+    title_margin: Optional[int] = None
     title_border_width: Optional[float] = None
     title_border_bottom: Optional[float] = None
     title_border_color: Optional[Color] = None
@@ -134,7 +136,7 @@ class ThemeJson(BaseModel):
     tab_margin: Optional[int] = None
     content_margin: Optional[Union[int, float, List[float]]] = None
     tab_dimming: bool = False
-    indicator_width: Optional[float] = None
+    indicator_width: Optional[int] = None   # matches ThemeSpec.indicator_width
     indicator_position: Optional[Union[str, List[str]]] = None
 
     # ---------------------------------------------------------------------------
